Moved MailDispatcherOptions to options folder
